requirements.txt
azure-cognitiveservices-speech==1.36.0
openai==0.28.0
pandas==1.3.5
pytz==2024.1
requests==2.31.0
Flask==1.1.2